Fixed installing hooks.

+++ .github/install.sh
@@ -13,18 +13,21 @@
 HERE=$(cd "$(dirname "$0")"; pwd -P)
 GIT_DIR=${HERE}/../.git
 LOCKFILE=${GIT_DIR}/.commit
+HOOKS="post-commit pre-commit prepare-commit-msg"
 
 GIT=$(command -v git)
+CHMOD=$(command -v chmod)
 CP=$(command -v cp)
 RM=$(command -v rm)
 
-if [ -e "${GIT_DIR}/hooks" ] && \
-   [ "" != "${GIT}" ] && [ "" != "${CP}" ] && [ "" != "${RM}" ]; \
+if [ -e "${GIT_DIR}/hooks" ] && [ "${GIT}" ] && \
+   [ "${CHMOD}" ] && [ "${CP}" ] && [ "${RM}" ]; \
 then
   # make sure the path to .gitconfig is a relative path
   ${GIT} config --local include.path ../.gitconfig 2>/dev/null
-  ${CP} "${HERE}/post-commit" "${GIT_DIR}/hooks"
-  ${CP} "${HERE}/pre-commit" "${GIT_DIR}/hooks"
-  ${CP} "${HERE}/prepare-commit-msg" "${GIT_DIR}/hooks"
+  for HOOK in ${HOOKS}; do
+    ${CP} "${HERE}/${HOOK}" "${GIT_DIR}/hooks"
+    ${CHMOD} +x "${GIT_DIR}/hooks/${HOOK}"
+  done
   ${RM} -f "${LOCKFILE}-version"
 fi
